# backend/public/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Vote Électronique",
        "description": "API pour le système de vote électronique",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur Local"
        }
    ],
    "paths": {
        "/api/audit/logs": {
            "get": {
                "tags": [
                    "Audit"
                ],
                "summary": "Lister les logs avec filtres",
                "operationId": "123f0c13eaea298dcd8e765e0a312df2",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Filtrer par ID utilisateur",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "action",
                        "in": "query",
                        "description": "Filtrer par action",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "model",
                        "in": "query",
                        "description": "Filtrer par modèle",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "days",
                        "in": "query",
                        "description": "Filtrer par nombre de jours",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des logs"
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/audit/logs/{id}": {
            "get": {
                "tags": [
                    "Audit"
                ],
                "summary": "Détails d'un log",
                "operationId": "77ea6db37d005bd1e1567b0d0ebbf500",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du log",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du log"
                    },
                    "404": {
                        "description": "Log non trouvé"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/audit/participations": {
            "get": {
                "tags": [
                    "Audit"
                ],
                "summary": "Lister les participations",
                "operationId": "13c8fb7ed4cb56b32445617ec70e8fe2",
                "parameters": [
                    {
                        "name": "election_id",
                        "in": "query",
                        "description": "Filtrer par élection",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Date de début",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Date de fin",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des participations"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/audit/elections/{id}/integrity": {
            "get": {
                "tags": [
                    "Audit"
                ],
                "summary": "Vérifier l'intégrité d'une élection",
                "operationId": "1288eb25ebe301a098602778dd35a368",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'élection",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Résultat de la vérification d'intégrité"
                    },
                    "404": {
                        "description": "Élection non trouvée"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/audit/stats": {
            "get": {
                "tags": [
                    "Audit"
                ],
                "summary": "Statistiques générales du système",
                "operationId": "e348da2c7410802a212dc929387788b6",
                "responses": {
                    "200": {
                        "description": "Statistiques du système"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/audit/export/logs": {
            "get": {
                "tags": [
                    "Audit"
                ],
                "summary": "Exporter les logs en CSV",
                "operationId": "134d60e9dda16b5c5da83cdc62b58db5",
                "parameters": [
                    {
                        "name": "days",
                        "in": "query",
                        "description": "Nombre de jours à exporter",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Filtrer par utilisateur",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche par modèle",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fichier CSV des logs"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/audit/export/report/{election_id}": {
            "get": {
                "tags": [
                    "Audit"
                ],
                "summary": "Exporter le rapport PDF d'une élection",
                "operationId": "ad3f4b233339693e4faa02f19bb891bb",
                "parameters": [
                    {
                        "name": "election_id",
                        "in": "path",
                        "description": "ID de l'élection",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fichier PDF du rapport"
                    },
                    "404": {
                        "description": "Élection non trouvée"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}